+++ src/Asset.php
@@ -2,17 +2,14 @@
 /**
  * PHP library for save CSS and JS files to be displayed in same place.
  * 
- * @category   JST
- * @package    Asset
- * @subpackage Asset
  * @author     volter9
  * @author     QsmaPL
+ * @author     Josantonius
  * @author     Josantonius - info@josantonius.com
  * @copyright  Copyright (c) 2016 JST PHP Framework
  * @license    https://opensource.org/licenses/MIT - The MIT License (MIT)
- * @version    1.1.0
  * @link       https://github.com/Josantonius/PHP-Asset
- * @since      File available since 1.0.0 - Update: 2017-01-30
+ * @since      File available since 1.0.0 - Update: 2017-02-17
  */
 
 namespace Josantonius\Asset;
@@ -34,7 +31,7 @@ class Asset {
      * @var array → asset templates
      */
     protected static $templates = [
-        'js'  => '<script src="%s" type="text/javascript"></script>',
+        'js'  => '<script%s src="%s" type="text/javascript"></script>',
         'css' => '<link href="%s" rel="stylesheet" type="text/css">',
     ];
 
@@ -45,21 +42,26 @@ class Asset {
      *
      * @param string|array $files    → file or files to add
      * @param string       $template → template type
+     * @param string       $attr     → attribute for loading JS files
      */
-    protected static function resource($files, $template) {
+    protected static function resource($files, $template, $attr = '') {
 
         $template = self::$templates[$template];
 
         if (is_array($files)) {
 
-            foreach ($files as $file) {
+            foreach ($files as $attribute => $file) {
 
-                echo sprintf($template, $file) . "\n";
+                $attr = (is_int($attribute)) ? '' : ' ' . $attribute;
+
+                echo sprintf($template, $attr, $file) . "\n";
             }
 
         } else {
 
-            echo sprintf($template, $files) . "\n";
+            $attr = (!empty($attr)) ? ' ' . $attr : '';
+
+            echo sprintf($template, $attr, $files) . "\n";
         }
     }
 
@@ -69,10 +71,11 @@ class Asset {
      * @since 1.0.0
      *
      * @param string|array $files → file or files to add
+     * @param string       $attr  → attribute for loading JS files
      */
-    public static function js($files) {
+    public static function js($files, $attr = '') {
 
-        static::resource($files, 'js');
+        static::resource($files, 'js', $attr);
     }
 
     /**